feat: uniform representation of entity repos
